Feat: 모임 페이지 생성 api 구현 성공

+++ travelmaker_android/app/src/main/java/com/gumibom/travelmaker/data/dto/request/MarkerCategoryPositionRequestDTO.kt
@@ -3,6 +3,6 @@ package com.gumibom.travelmaker.data.dto.request
 data class MarkerCategoryPositionRequestDTO(
     val latitude : Double,
     val longitude : Double,
-    val radius : Int,
+    val radius : Double,
     val categories : List<String>
 )
